[refactor] use the BlockSyncService as a component to fetch the blocks and commits from both the Observer and Validator services. Refactored the observer client request to align with the validator one.
