+++ consensus/core/src/lib.rs
@@ -8,6 +8,7 @@ mod authority_service;
 mod base_committer;
 mod block;
 mod block_manager;
+mod block_sync_service;
 mod block_verifier;
 mod commit;
 mod commit_consumer;
