+++ consensus/core/src/network/test_network.rs
@@ -184,6 +184,8 @@ impl ObserverNetworkService for Mutex<TestService> {
         &self,
         _peer: NodeId,
         _block_refs: Vec<BlockRef>,
+        _highest_accepted_rounds: Vec<Round>,
+        _breadth_first: bool,
     ) -> ConsensusResult<Vec<Bytes>> {
         unimplemented!("ObserverNetworkService fetch_blocks not implemented for TestService")
     }
